Legacy/project/requirements.txt
# FastAPI + Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
psycopg2-binary>=2.9.9

# HTTP
httpx>=0.25.0
requests>=2.31.0

# LLM / Parsing
python-dotenv>=1.0.0
google-genai>=0.8.0

# RSS + Scraping
feedparser>=6.0.0
beautifulsoup4>=4.12.0

# PDF Ingestion
pypdf>=3.17.0

# Validation
pydantic>=2.5.0
